+++ Deeploy/Targets/Generic/Layers.py
@@ -114,7 +114,7 @@ class RQSiHardswishLayer(iHardswishLayer):
         super().__init__(maps)
 
 
-class iSoftmaxLayer(ONNXLayer):
+class SoftmaxLayer(ONNXLayer):
 
     def __init__(self, maps: List[NodeMapper]):
         super().__init__(maps)
@@ -193,13 +193,13 @@ class RQMatMulLayer(MatMulLayer):
         return matmul + rqs
 
 
-class IntegerDivLayer(ONNXLayer):
+class DivLayer(ONNXLayer):
 
     def __init__(self, maps: List[NodeMapper]):
         super().__init__(maps)
 
 
-class RQIntegerDivLayer(IntegerDivLayer):
+class RQIntegerDivLayer(DivLayer):
 
     def __init__(self, maps: List[NodeMapper]):
         super().__init__(maps)
@@ -365,7 +365,13 @@ class ReduceSumLayer(ONNXLayer):
         return (inputShapes, outputShapes)
 
 
-class iLayerNormLayer(ONNXLayer):
+class ReluLayer(ONNXLayer):
+
+    def __init__(self, maps: List[NodeMapper]):
+        super().__init__(maps)
+
+
+class LayerNormLayer(ONNXLayer):
 
     def __init__(self, maps: List[NodeMapper]):
         super().__init__(maps)