+++ TargetLibraries/Generic/inc/kernel/Convolution.h
@@ -59,4 +59,10 @@ void Conv2d_s8_s8_s32_NCHW(int8_t const *__restrict__ pSrcA, uint32_t C,
                            int32_t *__restrict__ pDstC, int32_t input_offset,
                            int32_t output_offset);
 
+void Conv2d_fp32_fp32_fp32_NCHW(const float *__restrict__ pSrcA, uint32_t C,
+                                uint32_t H_padded, uint32_t W_padded,
+                                const float *__restrict__ pSrcB, uint32_t F,
+                                uint32_t P, uint32_t Q, uint32_t SP, uint32_t SQ,
+                                float *__restrict__ pDstC);
+
 #endif //__DEEPLOY_BASIC_MATH_CONVOLUTION_KERNEL_HEADER_